pyproject.toml
[project]
name = "delta-optimizer"
version = "0.1.0"
description = "Bot-maker research system for OA-buildable credit-spread strategies"
requires-python = ">=3.11,<3.13"
readme = "PROJECT_BRIEF.md"
authors = [{ name = "Aryan" }]

dependencies = [
    # Data layer
    "httpx>=0.27",
    "tenacity>=9.0",
    "diskcache>=5.6",
    "polars>=1.10",
    "pandas>=2.2",
    "duckdb>=1.1",
    "pyarrow>=17.0",
    "fredapi>=0.5",
    # Pricing / Greeks
    "py_vollib>=1.0.1",
    "QuantLib>=1.34",
    # Numerics + optimization
    "numpy>=1.26,<2.0",
    "scipy>=1.13",
    "optuna>=4.0",
    "skfolio>=0.5",
    "arch>=7.0",
    # CLI / UX
    "rich>=13.9",
    "typer>=0.13",
    "tqdm>=4.66",
    "python-dotenv>=1.0",
    # Misc
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "pytest-xdist>=3.6",
    "hypothesis>=6.115",
    "ruff>=0.7",
    "mypy>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/delta_optimizer"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]  # line length handled by formatter

[tool.mypy]
python_version = "3.11"
strict = true
files = ["src/delta_optimizer"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "slow: marks tests as slow (use -m 'not slow' to skip)",
    "network: marks tests requiring network access",
]

[tool.coverage.run]
source = ["src/delta_optimizer"]
branch = true
